revert: "feat(phai): llma-parser-recipe-create with server-side validation (#69287)"
This reverts commit 0528aedb6c2cfae5c30c4aa4ce5dda7e0f8a5e61.

Generated-By: PostHog Code
Task-Id: 2ec000a6-64ea-4f40-a9ac-e4f9a4133e59
