+++ products/ai_observability/frontend/components/EventContentWithAsyncData.tsx
@@ -105,7 +105,6 @@ export function EventContentConversation({
 
     const openCustomParserMax = useCustomParserMaxTool({
         eventId,
-        traceId,
         input,
         output,
         tools: loadedTools,
