+++ products/ai_observability/frontend/customParser/useCustomParserMaxTool.ts
@@ -6,8 +6,7 @@ import { lemonToast } from '@posthog/lemon-ui'
 import { useMaxTool } from 'scenes/max/useMaxTool'
 import { teamLogic } from 'scenes/teamLogic'
 
-import { useAttachedContext, useMcpToolApplyBack } from 'products/posthog_ai/frontend/api/logics'
-import type { AttachedContextItem } from 'products/posthog_ai/frontend/api/types'
+import { useAttachedContext } from 'products/posthog_ai/frontend/api/logics'
 
 import { llmAnalyticsParserRecipesCreate } from '../generated/api'
 import { parserRecipesLogic } from '../settings/parserRecipesLogic'
@@ -16,23 +15,8 @@ import { handleCreateParserRecipeCall } from './validateRecipe'
 
 const MAX_EXISTING_RECIPES_CONTEXT_LENGTH = 4000
 
-// Static, trusted instruction — never interpolate user or ingested trace data into it.
-// Points the headless agent at the reference + create tools and the trace_id/event_uuid it
-// finds in the untrusted trace-event sample block.
-const AI_TRACE_PARSER_TOOL_CONTEXT_ITEM: AttachedContextItem = {
-    type: 'instructions',
-    hidden: true,
-    value:
-        'The user has an LLM trace event open that fell back to raw JSON because no parser recognized it. ' +
-        'To fix it, first call llma-parser-recipe-reference for the recipe DSL syntax and examples, then call ' +
-        'llma-parser-recipe-create with the trace_id and event_uuid from the attached trace event sample. The ' +
-        'samples in context are truncated, but the server compiles and validates your recipe against the full ' +
-        'event and only saves it when it works. The open trace re-renders once the recipe is saved.',
-}
-
 export interface CustomParserMaxToolOptions {
     eventId: string
-    traceId?: string | null
     input: unknown
     output: unknown
     tools?: unknown
@@ -49,7 +33,6 @@ export interface CustomParserMaxToolOptions {
  */
 export function useCustomParserMaxTool({
     eventId,
-    traceId,
     input,
     output,
     tools,
@@ -122,13 +105,10 @@ export function useCustomParserMaxTool({
     useAttachedContext(
         active
             ? [
-                  AI_TRACE_PARSER_TOOL_CONTEXT_ITEM,
                   { type: 'llm_trace_event', key: eventId },
                   {
                       type: 'ai_trace_parser_context',
                       value: JSON.stringify({
-                          trace_id: traceId ?? null,
-                          event_uuid: eventId,
                           event_type: isGeneration ? 'generation' : 'span',
                           unrecognized,
                           sample_input: sampleForContext(input),
@@ -139,32 +119,6 @@ export function useCustomParserMaxTool({
               ]
             : null
     )
-
-    // Headless mirror: when the agent's llma-parser-recipe-create tool completes, reload the team's
-    // recipes so the live normalizer re-renders the open trace. loadRecipes is idempotent, so we run
-    // it on every completion regardless of whether the tool actually saved.
-    useMcpToolApplyBack({
-        tools: ['llma-parser-recipe-create'],
-        targetKey: 'ai-trace-parser',
-        active,
-        onApply: (event, { innerInput }) => {
-            loadRecipes()
-            // A create for a different event must not claim this one was fixed; when the
-            // args are unparseable we keep the toast, matching the pre-innerInput behavior.
-            if (innerInput && typeof innerInput.event_uuid === 'string' && innerInput.event_uuid !== eventId) {
-                return
-            }
-            try {
-                const output = event.invocation.output
-                const serialized = typeof output === 'string' ? output : JSON.stringify(output)
-                if (serialized && serialized.includes('recipe_id')) {
-                    lemonToast.success('Custom parser saved')
-                }
-            } catch {
-                // Output shape is client-profile-dependent — the agent narrates failures in chat.
-            }
-        },
-    })
 
     const { openMax } = useMaxTool({
         identifier: 'create_ai_trace_parser',
