+++ products/ai_observability/frontend/normalizer/index.ts
@@ -6,4 +6,3 @@ export type { RunOutcome } from './recipe/runtime/pipeline'
 export { mergeRecipes } from './recipe/merge'
 export type { StoredRecipe } from './recipe/merge'
 export { compileRecipe } from './recipe/compile/compiler'
-export { setNormalizerTelemetry } from './telemetry'
