+++ products/ai_observability/frontend/normalizer/recipe/recipeNormalizer.test.ts
@@ -1,4 +1,4 @@
-import { AVAILABLE_TOOLS_ROLE } from '../roles'
+import { AVAILABLE_TOOLS_ROLE } from '../../utils'
 import { RecipeNormalizer } from './recipeNormalizer'
 
 const mockRun = jest.fn()
