+++ products/ai_observability/frontend/normalizer/recipe/runtime/coercion.ts
@@ -1,6 +1,7 @@
+import posthog from 'posthog-js'
+
 import { CompatMessage, CompatToolCall, MultiModalContentItem } from '../../../types'
-import { normalizeRole } from '../../roles'
-import { emitNormalizerTelemetry } from '../../telemetry'
+import { normalizeRole } from '../../../utils'
 import { Expr } from '../ast/expr'
 import { Scope } from '../scope'
 import { EmitSpec, RoleTag } from '../spec/emitSpec'
@@ -117,7 +118,7 @@ export class SlotCoercer {
         // Reaching here is a recipe bug or coverage gap (recipes opt into rendering
         // arbitrary values with `stringify`); surface '' rather than plausible-looking
         // JSON, and log type only since the value may be sensitive.
-        emitNormalizerTelemetry('llma recipe produced non-text content', { content_type: typeof value })
+        posthog.capture('llma recipe produced non-text content', { content_type: typeof value })
         return ''
     }
