+++ products/ai_observability/frontend/normalizer/recipe/runtime/pipeline.test.ts
@@ -4,6 +4,8 @@ import { Rule } from '../ast/rule'
 import { Recipe } from '../spec/recipe'
 import { NO_MATCH, RecipePipeline } from './pipeline'
 
+jest.mock('posthog-js', () => ({ __esModule: true, default: { capture: jest.fn() } }))
+
 const MATCH_ALL = new Pattern({})
 const MATCH_NONE = new Pattern({ __absent: new ExistsPredicate(true) })
