+++ products/ai_observability/frontend/utils.ts
@@ -824,9 +824,30 @@ export function isOTelPartsMessage(input: unknown): input is OTelPartsMessage {
     )
 }
 
-// Role helpers live in the portable normalizer module (leaf, no browser deps);
-// re-exported here so the many frontend consumers keep their import paths.
-export { roleMap, normalizeRole, AVAILABLE_TOOLS_ROLE } from './normalizer/roles'
+export const roleMap: Record<string, string> = {
+    user: 'user',
+    human: 'user',
+
+    assistant: 'assistant',
+    model: 'assistant',
+    ai: 'assistant',
+    bot: 'assistant',
+
+    system: 'system',
+    instructions: 'system',
+    context: 'system',
+}
+
+export function normalizeRole(rawRole: unknown, fallback: string): string {
+    if (typeof rawRole !== 'string') {
+        return fallback
+    }
+    const lowercased = rawRole.toLowerCase()
+    return roleMap[lowercased] || lowercased
+}
+
+// Synthetic role used to surface the `$ai_tools` payload as a pseudo-message
+export const AVAILABLE_TOOLS_ROLE = 'available tools'
 
 export const INTERNAL_THINKING_ROLE = 'assistant (thinking)'
 export const INTERNAL_TOOL_RESULT_ROLE = 'assistant (tool result)'
