+++ services/mcp/package.json
@@ -50,7 +50,6 @@
         "@modelcontextprotocol/ext-apps": "^1.5.0",
         "@modelcontextprotocol/sdk": "^1.29.0",
         "@posthog/mcp-analytics": "npm:@posthog/mcp@0.5.0",
-        "@posthog/products-ai-observability": "workspace:*",
         "@posthog/quill": "workspace:*",
         "@posthog/quill-charts": "workspace:*",
         "@toon-format/toon": "^2.1.0",
