+++ services/mcp/scripts/copy-instructions.ts
@@ -29,12 +29,6 @@ const PROMPTS = [
         src: 'services/mcp/playbooks',
         dest: 'shared/playbooks',
     },
-    {
-        // Parser recipe DSL reference served by `llma-parser-recipe-reference`;
-        // the Python MaxTool reads the same file, keeping one source of truth.
-        src: 'products/ai_observability/backend/prompts/parser_recipe_examples.yaml',
-        dest: 'shared/parser_recipe_examples.yaml',
-    },
 ]
 
 /** `auditing-the-fleet` → `auditingTheFleet` (valid JS identifier). */
