+++ services/mcp/tests/hono/setup.ts
@@ -13,12 +13,9 @@ vi.mock('cloudflare:workers', () => ({
     },
 }))
 
-// Mock shared prompt files (generated by scripts/copy-instructions.ts, not checked in)
+// Mock shared guidelines (generated by scripts/copy-instructions.ts, not checked in)
 vi.mock('@shared/guidelines.md', () => ({
     default: '# Guidelines\nTest guidelines content.',
-}))
-vi.mock('@shared/parser_recipe_examples.yaml', () => ({
-    default: 'rules: []\n',
 }))
 
 vi.mock('@/resources/ui-apps', async (importOriginal) => {
