+++ services/mcp/vitest.hono.config.mts
@@ -1,10 +1,21 @@
 import tsconfigPaths from 'vite-tsconfig-paths'
 import { defineConfig } from 'vitest/config'
 
-import { textLoader } from './tests/vitest-text-loader'
-
 export default defineConfig({
-    plugins: [tsconfigPaths({ root: '.' }), textLoader],
+    plugins: [
+        tsconfigPaths({ root: '.' }),
+        {
+            name: 'text-loader',
+            transform(code, id) {
+                if (id.endsWith('.md') || id.endsWith('.html')) {
+                    return {
+                        code: `export default ${JSON.stringify(code)}`,
+                        map: null,
+                    }
+                }
+            },
+        },
+    ],
     test: {
         globals: true,
         environment: 'node',
