+++ services/mcp/wrangler.jsonc
@@ -53,7 +53,7 @@
     "rules": [
         {
             "type": "Text",
-            "globs": ["**/*.md", "**/*.yaml"],
+            "globs": ["**/*.md"],
             "fallthrough": true,
         },
     ],
